Configure the workspace pet without leaving the chat (#481)
Right-click -> Options on the pet navigated the SPA to /app/profile,
unloading the agent chat the user was in. The pet options body is now a
shared PetOptionsPanel rendered both in the profile section and in a modal
opened from the pet's context menu, so configuration happens in place. The
modal outlives the pet's own visibility so toggling the pet off from inside
it doesn't slam the dialog shut.
